refactor: 챗봇 팩토리 구조 분리
