+++ apps/qna/chatbot/factories/dtos.py
@@ -1,0 +1,15 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Message:
+    role: str
+    content: str
+
+
+@dataclass
+class Payload:
+    messages: list[Message]
+    model: str
+    stream: bool
+    temperature: float
