+++ apps/qna/chatbot/factories/payload.py
@@ -1,18 +1,4 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class Message:
-    role: str
-    content: str
-
-
-@dataclass
-class Payload:
-    messages: list[Message]
-    model: str
-    stream: bool
-    temperature: float
+from .dtos import Message, Payload
 
 
 class GroqFactory:
